Moving EventEmitter to executor package
