+++ server/executor/trace_poller.go
@@ -7,7 +7,6 @@ import (
 	"log"
 	"time"
 
-	"github.com/kubeshop/tracetest/server/executor/event"
 	"github.com/kubeshop/tracetest/server/executor/pollingprofile"
 	"github.com/kubeshop/tracetest/server/model"
 	"github.com/kubeshop/tracetest/server/subscription"
@@ -42,7 +41,7 @@ func NewTracePoller(
 	updater RunUpdater,
 	assertionRunner AssertionRunner,
 	subscriptionManager *subscription.Manager,
-	eventEmitter event.Emitter,
+	eventEmitter EventEmitter,
 ) PersistentTracePoller {
 	return tracePoller{
 		updater:             updater,
